+++ packages/create-analog/template-blog/src/app/app.component.ts
@@ -1,0 +1,30 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+
+@Component({
+  selector: 'app-root',
+  standalone: true,
+  imports: [RouterOutlet],
+  template: `
+    <nav>
+      <a href="/">Home</a>
+    </nav>
+    <router-outlet></router-outlet>
+  `,
+  styles: [
+    `
+      :host {
+        max-width: 1280px;
+        margin: 0 auto;
+        padding: 2rem;
+        text-align: center;
+      }
+
+      nav {
+        text-align: left;
+        padding: 0 0 2rem 0;
+      }
+    `,
+  ],
+})
+export class AppComponent {}