+++ packages/platform/src/lib/options.ts
@@ -30,4 +30,6 @@ export interface Options {
   nitro?: NitroConfig;
   apiPrefix?: string;
   jit?: boolean;
+  index?: string;
+  workspaceRoot?: string;
 }