+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## [0.2.22](https://github.com/analogjs/analog/compare/0.2.21...0.2.22) (2023-11-22)
+
+### Bug Fixes
+
+- **create-analog:** add blog files to cli template ([7b33473](https://github.com/analogjs/analog/commit/7b33473f4fec88c5979f441343d02b7bd12d9958))
+- **platform:** add missing `index` option for the platform plugin ([#765](https://github.com/analogjs/analog/issues/765)) ([9bd6609](https://github.com/analogjs/analog/commit/9bd6609e29ae26c44e0d8f46828fdb9f43b3237a))
+- **vite-plugin-angular:** add Vitest plugin to transform async/await in Angular packages ([#760](https://github.com/analogjs/analog/issues/760)) ([8024c49](https://github.com/analogjs/analog/commit/8024c49cdb22e6e9f261f509cec86201581b14b8))
+
+### Features
+
+- **vite-plugin-nitro:** add `workspaceRoot` option to override `process.cwd` ([#764](https://github.com/analogjs/analog/issues/764)) ([90cac1c](https://github.com/analogjs/analog/commit/90cac1c84cc39fb7f28821529b8133f63016d9cc))
+
+## [0.2.21](https://github.com/analogjs/analog/compare/0.2.20...0.2.21) (2023-11-20)
+
+### Bug Fixes
+
+- **nx-plugin:** add angular-v17 template files ([#758](https://github.com/analogjs/analog/issues/758)) ([a0668d2](https://github.com/analogjs/analog/commit/a0668d29b4b2deb5c62851d0a79f4ab15ddece45))
+
+### Features
+
+- **create-analog:** add blog template to cli ([#761](https://github.com/analogjs/analog/issues/761)) ([f2c42df](https://github.com/analogjs/analog/commit/f2c42dff408e4ad5613dbc3ba49e40dfc79b5c71))
+- upgrade to Vite 5 and update template apps to Vite 5 ([#759](https://github.com/analogjs/analog/issues/759)) ([fa982dd](https://github.com/analogjs/analog/commit/fa982dd6987e7677562dff5e5a101e7502eeedf7))
+
+## [0.2.20](https://github.com/analogjs/analog/compare/0.2.19...0.2.20) (2023-11-15)
+
+### Bug Fixes
+
+- **nx-plugin:** add @angular/platform-server/init to main.server template files ([#747](https://github.com/analogjs/analog/issues/747)) ([f45a4df](https://github.com/analogjs/analog/commit/f45a4df3fa71cf1df9737a323b16762602e6ccdc))
+
 ## [0.2.19](https://github.com/analogjs/analog/compare/0.2.18...0.2.19) (2023-11-08)
 
 ### Bug Fixes